Move Branding-specific components to route folder
