+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/branding/components/options.ts
